requirements.txt
tensorflow
biopython
scikit-learn
numpy
matplotlib
pandas